perf: define the mercator world mapping next to the mercator functions it calls
